+++ yak-security/src/main/java/io/yak/framework/security/service/ProjectService.java
@@ -17,44 +17,44 @@ public interface ProjectService {
   public ProjectVO createProject(ProjectSaveDTO var1, String var2)
       throws YakSecurityException;
 
-  public ProjectVO getProjectDetailByProjectId(Integer var1)
+  public ProjectVO getProjectDetailByProjectId(Long var1)
       throws YakSecurityException;
 
-  public ProjectBriefVO getProjectBriefByProjectId(Integer var1);
+  public ProjectBriefVO getProjectBriefByProjectId(Long var1);
 
   public PagingData<ProjectVO> getProjectPage(ProjectQueryDTO var1);
 
   public PagingData<ProjectVO> getProjectPage(ProjectQueryDTO var1,
                                               List<Long> var2);
 
-  public void deleteProjectByProjectId(Integer var1, String var2);
+  public void deleteProjectByProjectId(Long var1, String var2);
 
   public void updateProject(ProjectSaveDTO var1, String var2)
       throws YakSecurityException;
 
-  public void changeProjectStatus(Integer var1, String var2);
+  public void changeProjectStatus(Long var1, String var2);
 
-  public void addProjectUser(Integer var1, Integer var2, String var3);
+  public void addProjectUser(Long var1, Long var2, String var3);
 
-  public void delProjectUser(Integer var1, Integer var2, String var3);
+  public void delProjectUser(Long var1, Long var2, String var3);
 
-  public void addProjectOwner(Integer var1, Integer var2, String var3);
+  public void addProjectOwner(Long var1, Long var2, String var3);
 
-  public void delProjectOwner(Integer var1, Integer var2, String var3);
+  public void delProjectOwner(Long var1, Long var2, String var3);
 
   public List<ProjectBriefVO> getProjectBriefList();
 
-  public ProjectDeleteCheckVO checkBeforeDelete(Integer var1);
+  public ProjectDeleteCheckVO checkBeforeDelete(Long var1);
 
   public PagingData<ProjectBriefVO>
   getProjectBriefPage(ProjectBriefQueryDTO var1);
 
-  public boolean checkProjectExist(Integer var1);
+  public boolean checkProjectExist(Long var1);
 
-  public Result<List<UserBriefVO>> unassignedByProjectId(Integer var1)
+  public Result<List<UserBriefVO>> unassignedByProjectId(Long var1)
       throws YakSecurityException;
 
-  public Result<List<ProjectBriefVO>> getProjectBriefByUserId(Integer var1);
+  public Result<List<ProjectBriefVO>> getProjectBriefByUserId(Long var1);
 
   public List<ProjectBriefVOWithUser>
   listProjectBriefVOWithUserByProjectIds(List<Long> var1);
